♻️ refactor: SCHOOLS_BY_BRANCH 상수를 shared/config으로 이동
shared/lib/formatSchoolName이 features 레이어를 역참조하는 구조 해소.
SCHOOLS_BY_BRANCH를 src/shared/config/schools.ts로 이동하고
projectFilterOptions에서 re-export하여 기존 소비처 영향 없음.

(리뷰 반영 #3254325836)
