+++ src/features/project/list/model/projectFilterOptions.ts
@@ -12,14 +12,7 @@ export const BRANCH_OPTIONS: ProjectFilterOption[] = [
   { value: "Xenon", label: "Xenon" },
 ]
 
-export const SCHOOLS_BY_BRANCH = {
-  Chromium: ["광운대", "동양미래대", "서울여대", "한국공학대", "한국외대"],
-  Ferrum: ["동국대", "이화여대", "홍익대 서울", "홍익대 세종"],
-  Neon: ["가천대", "동덕여대", "숙명여대", "인하대", "한국항공대"],
-  Platinum: ["동아대", "영남대", "인제대"],
-  Selenium: ["서경대", "성신여대", "숭실대", "안양대", "한양대 ERICA"],
-  Xenon: ["가톨릭대", "단국대", "덕성여대", "중앙대", "한성대"],
-} as const
+export { SCHOOLS_BY_BRANCH } from "@/shared/config/schools"
 
 export const PART_OPTIONS: ProjectFilterOption[] = [
   { value: "Web", label: "Web" },
